requirements.txt
langchain==0.3.9
langchain-core==0.3.47
langchain-community==0.3.9
langchain-groq==0.3.0
langchain-text-splitters==0.3.3
faiss-cpu==1.12.0
sentence-transformers==3.0.1
python-dotenv==1.0.1
PyPDF2==3.0.1
